Fix: Adjust Spots table, price column to use floating point to prevent Sequelize type coercion

+++ apps/backend/db/migrations/20240812031753-create-spot.js
@@ -53,7 +53,7 @@ module.exports = {
 					allowNull: false,
 				},
 				price: {
-					type: Sequelize.DECIMAL(10, 2),
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				createdAt: {
